extracting FileEmbeddedDocumentExtractor
